# Optimize performance of `string::ascii` function