--- a/datafusion/functions/benches/ascii.rs
+++ b/datafusion/functions/benches/ascii.rs
@@ -29,87 +29,105 @@ fn criterion_benchmark(c: &mut Criterion) {
     // All benches are single batch run with 8192 rows
     const N_ROWS: usize = 8192;
     const STR_LEN: usize = 16;
-    const NULL_DENSITY: f32 = 0.2;
     const UTF8_DENSITY_OF_ALL_ASCII: f32 = 0.0;
     const NORMAL_UTF8_DENSITY: f32 = 0.8;
 
-    // StringArray ASCII only
-    let args_string_ascii = gen_string_array(
-        N_ROWS,
-        STR_LEN,
-        NULL_DENSITY,
-        UTF8_DENSITY_OF_ALL_ASCII,
-        false,
-    );
-    c.bench_function("ascii/string_ascii_only", |b| {
-        b.iter(|| {
-            black_box(ascii.invoke_with_args(ScalarFunctionArgs {
-                args: args_string_ascii.clone(),
-                arg_fields: vec![&Field::new(
-                    "a",
-                    args_string_ascii[0].data_type(),
-                    true,
-                )],
-                number_rows: N_ROWS,
-                return_field: &Field::new("f", DataType::Utf8, true),
-            }))
-        })
-    });
+    for null_density in [0.0, 0.5] {
+        // StringArray ASCII only
+        let args_string_ascii = gen_string_array(
+            N_ROWS,
+            STR_LEN,
+            null_density,
+            UTF8_DENSITY_OF_ALL_ASCII,
+            false,
+        );
+        c.bench_function(
+            format!("ascii/string_ascii_only (null_density={null_density})").as_str(),
+            |b| {
+                b.iter(|| {
+                    black_box(ascii.invoke_with_args(ScalarFunctionArgs {
+                        args: args_string_ascii.clone(),
+                        arg_fields: vec![&Field::new(
+                            "a",
+                            args_string_ascii[0].data_type(),
+                            true,
+                        )],
+                        number_rows: N_ROWS,
+                        return_field: &Field::new("f", DataType::Utf8, true),
+                    }))
+                })
+            },
+        );
 
-    // StringArray UTF8
-    let args_string_utf8 =
-        gen_string_array(N_ROWS, STR_LEN, NULL_DENSITY, NORMAL_UTF8_DENSITY, false);
-    c.bench_function("ascii/string_utf8", |b| {
-        b.iter(|| {
-            black_box(ascii.invoke_with_args(ScalarFunctionArgs {
-                args: args_string_utf8.clone(),
-                arg_fields: vec![&Field::new("a", args_string_utf8[0].data_type(), true)],
-                number_rows: N_ROWS,
-                return_field: &Field::new("f", DataType::Utf8, true),
-            }))
-        })
-    });
+        // StringArray UTF8
+        let args_string_utf8 =
+            gen_string_array(N_ROWS, STR_LEN, null_density, NORMAL_UTF8_DENSITY, false);
+        c.bench_function(
+            format!("ascii/string_utf8 (null_density={null_density})").as_str(),
+            |b| {
+                b.iter(|| {
+                    black_box(ascii.invoke_with_args(ScalarFunctionArgs {
+                        args: args_string_utf8.clone(),
+                        arg_fields: vec![&Field::new(
+                            "a",
+                            args_string_utf8[0].data_type(),
+                            true,
+                        )],
+                        number_rows: N_ROWS,
+                        return_field: &Field::new("f", DataType::Utf8, true),
+                    }))
+                })
+            },
+        );
 
-    // StringViewArray ASCII only
-    let args_string_view_ascii = gen_string_array(
-        N_ROWS,
-        STR_LEN,
-        NULL_DENSITY,
-        UTF8_DENSITY_OF_ALL_ASCII,
-        true,
-    );
-    c.bench_function("ascii/string_view_ascii_only", |b| {
-        b.iter(|| {
-            black_box(ascii.invoke_with_args(ScalarFunctionArgs {
-                args: args_string_view_ascii.clone(),
-                arg_fields: vec![&Field::new(
-                    "a",
-                    args_string_view_ascii[0].data_type(),
-                    true,
-                )],
-                number_rows: N_ROWS,
-                return_field: &Field::new("f", DataType::Utf8, true),
-            }))
-        })
-    });
+        // StringViewArray ASCII only
+        let args_string_view_ascii = gen_string_array(
+            N_ROWS,
+            STR_LEN,
+            null_density,
+            UTF8_DENSITY_OF_ALL_ASCII,
+            true,
+        );
+        c.bench_function(
+            format!("ascii/string_view_ascii_only (null_density={null_density})")
+                .as_str(),
+            |b| {
+                b.iter(|| {
+                    black_box(ascii.invoke_with_args(ScalarFunctionArgs {
+                        args: args_string_view_ascii.clone(),
+                        arg_fields: vec![&Field::new(
+                            "a",
+                            args_string_view_ascii[0].data_type(),
+                            true,
+                        )],
+                        number_rows: N_ROWS,
+                        return_field: &Field::new("f", DataType::Utf8, true),
+                    }))
+                })
+            },
+        );
 
-    // StringViewArray UTF8
-    let args_string_view_utf8 =
-        gen_string_array(N_ROWS, STR_LEN, NULL_DENSITY, NORMAL_UTF8_DENSITY, true);
-    c.bench_function("ascii/string_view_utf8", |b| {
-        b.iter(|| {
-            black_box(ascii.invoke_with_args(ScalarFunctionArgs {
-                args: args_string_view_utf8.clone(),
-                arg_fields: vec![&Field::new(
-                    "a",
-                    args_string_view_utf8[0].data_type(),
-                    true,
-                )],
-                number_rows: N_ROWS,
-                return_field: &Field::new("f", DataType::Utf8, true),
-            }))
-        })
-    });
+        // StringViewArray UTF8
+        let args_string_view_utf8 =
+            gen_string_array(N_ROWS, STR_LEN, null_density, NORMAL_UTF8_DENSITY, true);
+        c.bench_function(
+            format!("ascii/string_view_utf8 (null_density={null_density})").as_str(),
+            |b| {
+                b.iter(|| {
+                    black_box(ascii.invoke_with_args(ScalarFunctionArgs {
+                        args: args_string_view_utf8.clone(),
+                        arg_fields: vec![&Field::new(
+                            "a",
+                            args_string_view_utf8[0].data_type(),
+                            true,
+                        )],
+                        number_rows: N_ROWS,
+                        return_field: &Field::new("f", DataType::Utf8, true),
+                    }))
+                })
+            },
+        );
+    }
 }
 
 criterion_group!(benches, criterion_benchmark);