@alamb 2025-05-21
This is better than what is on main and we can do always optimize further.

Thanks a log @tlm365 @findepi and @Dandandan  -- always up and to the right